requirements.txt
fhir.resources==6.5.0
fastapi==0.89.0
prometheus-fastapi-instrumentator==5.7.1
structlog==21.1.0
averbis-python-api==0.4.0
tenacity==8.0.1
aiokafka==0.7.2
